issue137 Remove package org.eclipse.microprofile.lra.client from the API
