+++ tck/src/main/java/org/eclipse/microprofile/lra/tck/participant/api/IllegalLRAStateException.java
@@ -17,40 +17,33 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package org.eclipse.microprofile.lra.client;
+package org.eclipse.microprofile.lra.tck.participant.api;
 
 import javax.ws.rs.WebApplicationException;
 
-public class InvalidLRAIdException extends WebApplicationException {
-    private static final long serialVersionUID = 1271422145863321852L;
+public class IllegalLRAStateException extends WebApplicationException {
     private final String lraId;
+    private final String operation;
 
     /**
-     * Invalid LRA id exception.
+     * Creation of LRA state exception.
      *
      * @param lraId  LRA id that is behind this exception
-     * @param message  error message of this exception
-     * @param cause  cause exception
-     */
-    public InvalidLRAIdException(String lraId, String message, Throwable cause) {
-        super(String.format("%s, LRA id: %s", message, lraId), cause);
-
-        this.lraId = lraId;
-    }
-
-    /**
-     * Invalid LRA id exception.
-     *
-     * @param lraId  LRA id that is behind this exception
+     * @param operation the java method that generated the exception
      * @param message  error message of this exception
      */
-    public InvalidLRAIdException(String lraId, String message) {
-        super(String.format("%s, LRA id: %s", message, lraId));
+    public IllegalLRAStateException(String lraId, String operation, String message) {
+        super(String.format("%s, lra id: %s", message, lraId));
 
         this.lraId = lraId;
+        this.operation = operation;
     }
 
     public String getLraId() {
         return this.lraId;
+    }
+
+    public String getOperation() {
+        return this.operation;
     }
 }
